backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.6
django-debug-toolbar==4.4.2
django-filter==24.2
django-rest-swagger==2.2.0
django-split-settings==1.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Markdown==3.6
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.1
psycopg2-binary==2.9.9
pymemcache==4.0.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
setuptools==70.0.0
simplejson==3.19.2
soupsieve==2.5
sqlparse==0.5.0
swagger-spec-validator==3.0.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
wheel==0.43.0
